+++ ref-impl/src/com-api/com-dll/CAAFInProcServer.h
@@ -1,6 +1,6 @@
 //=---------------------------------------------------------------------=
 //
-// $Id: CAAFInProcServer.h,v 1.11 2004/02/27 14:26:46 stuart_hc Exp $ $Name:  $
+// $Id: CAAFInProcServer.h,v 1.12 2004/10/25 13:25:31 stuart_hc Exp $ $Name:  $
 //
 // The contents of this file are subject to the AAF SDK Public
 // Source License Agreement (the "License"); You may not use this file
@@ -98,7 +98,7 @@ public:
 	const char* GetServerDirectory() const;
 
 protected:
-	long GetRegisterIndex(long index);
+	int GetRegisterIndex(int index);
 
 protected:
 	AAFComObjectInfo_t *_pObjectInfo;
